// Project 11.7/src/types/weather.ts
export namespace weather {
    export class City extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            code?: string;
            name?: string;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("code" in data && data.code != undefined) {
                    this.code = data.code;
                }
                if ("name" in data && data.name != undefined) {
                    this.name = data.name;
                }
            }
        }
        get code() {
            return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
        }
        set code(value: string) {
            pb_1.Message.setField(this, 1, value);
        }
        get name() {
            return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
        }
        set name(value: string) {
            pb_1.Message.setField(this, 2, value);
        }
        static fromObject(data: {
            code?: string;
            name?: string;
        }): City {
            const message = new City({});
            if (data.code != null) {
                message.code = data.code;
            }
            if (data.name != null) {
                message.name = data.name;
            }
            return message;
        }
        toObject() {
            const data: {
                code?: string;
                name?: string;
            } = {};
            if (this.code != null) {
                data.code = this.code;
            }
            if (this.name != null) {
                data.name = this.name;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.code.length)
                writer.writeString(1, this.code);
            if (this.name.length)
                writer.writeString(2, this.name);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): City {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new City();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.code = reader.readString();
                        break;
                    case 2:
                        message.name = reader.readString();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): City {
            return City.deserialize(bytes);
        }
    }
// ...
    export namespace CityQuery {
        export class Result extends pb_1.Message {
            #one_of_decls: number[][] = [];
            constructor(data?: any[] | {
                cities?: City[];
            }) {
                super();
                pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [2], this.#one_of_decls);
                if (!Array.isArray(data) && typeof data == "object") {
                    if ("cities" in data && data.cities != undefined) {
                        this.cities = data.cities;
                    }
                }
            }
            get cities() {
                return pb_1.Message.getRepeatedWrapperField(this, City, 2) as City[];
            }
            set cities(value: City[]) {
                pb_1.Message.setRepeatedWrapperField(this, 2, value);
            }
            static fromObject(data: {
                cities?: ReturnType<typeof City.prototype.toObject>[];
            }): Result {
                const message = new Result({});
                if (data.cities != null) {
                    message.cities = data.cities.map(item => City.fromObject(item));
                }
                return message;
            }
            toObject() {
                const data: {
                    cities?: ReturnType<typeof City.prototype.toObject>[];
                } = {};
                if (this.cities != null) {
                    data.cities = this.cities.map((item: City) => item.toObject());
                }
                return data;
            }
            serialize(): Uint8Array;
            serialize(w: pb_1.BinaryWriter): void;
            serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
                const writer = w || new pb_1.BinaryWriter();
                if (this.cities.length)
                    writer.writeRepeatedMessage(2, this.cities, (item: City) => item.serialize(writer));
                if (!w)
                    return writer.getResultBuffer();
            }
            static deserialize(bytes: Uint8Array | pb_1.BinaryReader): Result {
                const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new Result();
                while (reader.nextField()) {
                    if (reader.isEndGroup())
                        break;
                    switch (reader.getFieldNumber()) {
                        case 2:
                            reader.readMessage(message.cities, () => pb_1.Message.addToRepeatedWrapperField(message, 2, City.deserialize(reader), City));
                            break;
                        default: reader.skipField();
                    }
                }
                return message;
            }
            serializeBinary(): Uint8Array {
                return this.serialize();
            }
            static deserializeBinary(bytes: Uint8Array): Result {
                return Result.deserialize(bytes);
            }
        }
    }
// ...
    export namespace Forecast {
        export class Result extends pb_1.Message {
            #one_of_decls: number[][] = [];
            constructor(data?: any[] | {
                temperature?: Temperature;
            }) {
                super();
                pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
                if (!Array.isArray(data) && typeof data == "object") {
                    if ("temperature" in data && data.temperature != undefined) {
                        this.temperature = data.temperature;
                    }
                }
            }
            get temperature() {
                return pb_1.Message.getWrapperField(this, Temperature, 1) as Temperature;
            }
            set temperature(value: Temperature) {
                pb_1.Message.setWrapperField(this, 1, value);
            }
            get has_temperature() {
                return pb_1.Message.getField(this, 1) != null;
            }
            static fromObject(data: {
                temperature?: ReturnType<typeof Temperature.prototype.toObject>;
            }): Result {
                const message = new Result({});
                if (data.temperature != null) {
                    message.temperature = Temperature.fromObject(data.temperature);
                }
                return message;
            }
            toObject() {
                const data: {
                    temperature?: ReturnType<typeof Temperature.prototype.toObject>;
                } = {};
                if (this.temperature != null) {
                    data.temperature = this.temperature.toObject();
                }
                return data;
            }
            serialize(): Uint8Array;
            serialize(w: pb_1.BinaryWriter): void;
            serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
                const writer = w || new pb_1.BinaryWriter();
                if (this.has_temperature)
                    writer.writeMessage(1, this.temperature, () => this.temperature.serialize(writer));
                if (!w)
                    return writer.getResultBuffer();
            }
            static deserialize(bytes: Uint8Array | pb_1.BinaryReader): Result {
                const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new Result();
                while (reader.nextField()) {
                    if (reader.isEndGroup())
                        break;
                    switch (reader.getFieldNumber()) {
                        case 1:
                            reader.readMessage(message.temperature, () => message.temperature = Temperature.deserialize(reader));
                            break;
                        default: reader.skipField();
                    }
                }
                return message;
            }
            serializeBinary(): Uint8Array {
                return this.serialize();
            }
            static deserializeBinary(bytes: Uint8Array): Result {
                return Result.deserialize(bytes);
            }
        }
    }
// ...
}